Extract document sync and events
