+++ core/events.py
@@ -1,0 +1,28 @@
+try:
+    from typing import Any, List, Dict, Tuple, Callable, Optional
+    assert Any and List and Dict and Tuple and Callable and Optional
+except ImportError:
+    pass
+
+
+class Events:
+    listener_dict = dict()  # type: Dict[str, Callable[..., None]]
+
+    @classmethod
+    def subscribe(cls, key, listener):
+        if key in cls.listener_dict:
+            cls.listener_dict[key].append(listener)
+        else:
+            cls.listener_dict[key] = [listener]
+        return lambda: cls.unsubscribe(key, listener)
+
+    @classmethod
+    def unsubscribe(cls, key, listener):
+        if key in cls.listener_dict:
+            cls.listener_dict[key].remove(listener)
+
+    @classmethod
+    def publish(cls, key, *args):
+        if key in cls.listener_dict:
+            for listener in cls.listener_dict[key]:
+                listener(*args)
